% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.asDate.R
\name{ds.asDate}
\alias{ds.asDate}
\title{Converts a server-side R object into a Date class}
\usage{
ds.asDate(
  x.name,
  format = NULL,
  origin = NULL,
  newobj = NULL,
  add_as_column = TRUE,
  datasources = NULL
)
}
\arguments{
\item{x.name}{a character string providing the name of the input object to be coerced to date character.}

\item{format}{A character string. The default for the format methods is "\%Y-\%m-\%d".
Check the details section of \link{strptime} for further information.}

\item{origin}{something which can be coerced by \code{as_Date(origin)} to such an object (default: the Unix epoch of "1970-01-01").
Note that in this instance, the server values of \code{x.name} is assumed to reflect the number of days since \code{origin}.}

\item{newobj}{a character string that provides the name for the output object that is stored on the data servers.
Default \code{"asDate.newobj"}.}

\item{add_as_column}{A logic value to indicate if the converted values are to be added as a column to the input dataframe.
This option is ignored if \code{x.name} is not a dataframe. The column name will be \code{newobj}. Defaults to \code{TRUE}.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Does not have a return object.
}
\description{
Converts the input object into a date class. This function is based on the lubridate function \code{as_date}
}
