% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.dateDiff.R
\name{ds.dateDiff}
\alias{ds.dateDiff}
\title{Computes diference in days between two date objects}
\usage{
ds.dateDiff(
  time1,
  time2,
  newobj = NULL,
  add_as_column = TRUE,
  datasources = NULL
)
}
\arguments{
\item{time1}{a character string providing the name of a Date object on the server}

\item{time2}{a character string providing the name of a Date object on the server}

\item{newobj}{a character string that provides the name for the output object that is stored on the data servers.
Default \code{"dateDiff"}.}

\item{add_as_column}{A logic value to indicate if the converted values are to be added as a column to the input dataframe.
This option is ignored if \code{x.name} is not a dataframe. The column name will be \code{newobj}. Defaults to \code{TRUE}.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Does not have a return object.
}
\description{
Computes the time difference using the lubridate package. Both input objects have to be created with \link{ds.asDate}.
}
\details{
The time difference is computed as an absolute value.
}
