% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.subsetByDate.R
\name{ds.subsetByDate}
\alias{ds.subsetByDate}
\title{Subsets data frames in the server side}
\usage{
ds.subsetByDate(
  x.name,
  date_column,
  day = NULL,
  month = NULL,
  year = NULL,
  range = NULL,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{x.name}{a character string providing the name of the dataframe to be subsetted.}

\item{date_column}{a character specifying the column of \code{x.name} of type Date to be used for the subset computation.}

\item{day}{a numeric value to specify a day to subset}

\item{month}{a numeric value to specify a month to subset}

\item{year}{a numeric value to specify a year to subset}

\item{range}{a character vector (of length 2) to provide a start and end date to perform the subset. Use the format \code{"yyyy-mm-dd"}.}

\item{newobj}{a character string that provides the name for the output object that is stored on the data servers.
Default \code{"dateSubset"}.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Does not have a return object.
}
\description{
Subsets a dataframe by concrete dates or ranges.
}
\details{
The \code{day}, \code{month} and \code{year} arguments can be combined (e.g. to subset the year 2022 month 1).
}
