% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.dateDiff.R
\name{ds.dateDiff}
\alias{ds.dateDiff}
\title{Computes difference between two date objects}
\usage{
ds.dateDiff(
  time1,
  time2,
  unit = "days",
  as_integer = FALSE,
  newobj = NULL,
  add_as_column = TRUE,
  datasources = NULL
)
}
\arguments{
\item{time1}{a character string providing the name of a Date object on the server}

\item{time2}{a character string providing the name of a Date object on the server}

\item{unit}{a character string specifying the unit for the difference: "days" (default), "months", or "years"}

\item{as_integer}{a logical value indicating whether to truncate the result to integer values (remove decimals).
Default is \code{FALSE} (keeps decimal values).}

\item{newobj}{a character string that provides the name for the output object that is stored on the data servers.
Default \code{"dateDiff"}.}

\item{add_as_column}{A logic value to indicate if the converted values are to be added as a column to the input dataframe.
This option is ignored if \code{x.name} is not a dataframe. The column name will be \code{newobj}. Defaults to \code{TRUE}.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Does not have a return object.
}
\description{
Computes the time difference between two Date objects. Both input objects have to be created with \link{ds.asDate} or \link{ds.makeDate}.
The difference can be computed in days (default), months, or years.
}
\details{
The time difference is computed as an absolute value. When using months or years, the result is calculated by dividing
the difference in days by the average number of days per month (30.44) or per year (365.25), respectively.
}
\examples{
\dontrun{
# Difference in days (default)
ds.dateDiff(time1 = "birth_date", time2 = "visit_date", unit = "days")

# Difference in months
ds.dateDiff(time1 = "birth_date", time2 = "visit_date", unit = "months")

# Difference in years
ds.dateDiff(time1 = "birth_date", time2 = "visit_date", unit = "years")

# Difference in years as integer (truncated)
ds.dateDiff(time1 = "birth_date", time2 = "visit_date", unit = "years", as_integer = TRUE)
}

}
